+++ libsolidity/codegen/ABIFunctions.h
@@ -222,6 +222,8 @@ private:
 	std::string abiDecodingFunctionCalldataArray(ArrayType const& _type);
 	/// Part of @a abiDecodingFunction for byte array types.
 	std::string abiDecodingFunctionByteArray(ArrayType const& _type, bool _fromMemory);
+	/// Part of @a abiDecodingFunction for calldata struct types.
+	std::string abiDecodingFunctionCalldataStruct(StructType const& _type);
 	/// Part of @a abiDecodingFunction for struct types.
 	std::string abiDecodingFunctionStruct(StructType const& _type, bool _fromMemory);
 	/// Part of @a abiDecodingFunction for array types.